services/crawler-service/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
requests==2.31.0
beautifulsoup4==4.12.2
aiohttp==3.8.5
asyncio==3.4.3
httpx==0.25.0
lxml==4.9.3
pydantic==2.4.2
python-dotenv==1.0.0
googlesearch-python==1.2.3
nltk==3.8.1
Scrapy==2.11.0
pymongo==4.6.0
elasticsearch==8.10.0
redis==5.0.1 